requirements.txt
flask
flask-cors
flask-bcrypt
psycopg2-binary
flask-script
pandas